requirements.txt
smbus
pyalsaaudio